Recombine media serializers in an organised file
