+++ api/catalog/api/serializers/audio_serializers.py
@@ -2,14 +2,83 @@ from rest_framework import serializers
 
 from elasticsearch_dsl.response import Hit
 
+from catalog.api.constants.field_values import AUDIO_CATEGORIES, DURATION
 from catalog.api.docs.media_docs import fields_to_md
-from catalog.api.models import Audio, AudioSet
-from catalog.api.serializers.fields import SchemableHyperlinkedIdentityField
-from catalog.api.serializers.response.media import (
+from catalog.api.models import Audio, AudioReport, AudioSet
+from catalog.api.serializers.fields import (
+    EnumCharField,
+    SchemableHyperlinkedIdentityField,
+)
+from catalog.api.serializers.media_serializers import (
+    MediaSearchRequestSerializer,
     MediaSearchSerializer,
     MediaSerializer,
     get_hyperlinks_serializer,
+    get_search_request_source_serializer,
 )
+
+
+#######################
+# Request serializers #
+#######################
+
+
+AudioSearchRequestSourceSerializer = get_search_request_source_serializer("audio")
+
+
+class AudioSearchRequestSerializer(
+    AudioSearchRequestSourceSerializer,
+    MediaSearchRequestSerializer,
+):
+    """Parse and validate search query string parameters."""
+
+    fields_names = [
+        *MediaSearchRequestSerializer.fields_names,
+        *AudioSearchRequestSourceSerializer.field_names,
+        "category",
+        "duration",
+    ]
+    """
+    Keep the fields names in sync with the actual fields below as this list is
+    used to generate Swagger documentation.
+    """
+
+    category = EnumCharField(
+        plural="categories",
+        enum_var=AUDIO_CATEGORIES,
+        required=False,
+    )
+    duration = EnumCharField(
+        plural="durations",
+        enum_var=DURATION,
+        required=False,
+    )
+
+
+class AudioReportSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = AudioReport
+        fields = ["identifier", "reason", "description"]
+        read_only_fields = ["identifier"]
+
+    def create(self, validated_data):
+        if (
+            validated_data["reason"] == "other"
+            and (
+                "description" not in validated_data
+                or len(validated_data["description"])
+            )
+            < 20
+        ):
+            raise serializers.ValidationError(
+                "Description must be at least be 20 characters long"
+            )
+        return AudioReport.objects.create(**validated_data)
+
+
+########################
+# Response serializers #
+########################
 
 
 class AudioSetSerializer(serializers.ModelSerializer):
@@ -28,7 +97,7 @@ class AudioSetSerializer(serializers.ModelSerializer):
         ]
 
 
-AudioHyperlinksSerializer = get_hyperlinks_serializer("audio")  # class
+AudioHyperlinksSerializer = get_hyperlinks_serializer("audio")
 
 
 class AudioSerializer(AudioHyperlinksSerializer, MediaSerializer):
@@ -92,6 +161,11 @@ class AudioSearchSerializer(MediaSearchSerializer):
             f"{fields_to_md(AudioSerializer.Meta.fields)}."
         ),
     )
+
+
+##########################
+# Additional serializers #
+##########################
 
 
 class AudioWaveformSerializer(serializers.Serializer):
